+++ tests/srcs/history/bang.c
@@ -9,7 +9,7 @@
 #define NB_SUB_TESTS 38
 #define NB_ERR_TESTS 5
 
-// #define BANG_TEST_VERBOSE
+#define BANG_TEST_VERBOSE
 
 /*
 	to test your own strings:
